% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_biplot.R
\name{filtered_biplot}
\alias{filtered_biplot}
\title{produce a biplot based on sharing of information between a `prcomp`-like output and a SingleCellExperiment}
\usage{
filtered_biplot(
  prcomp_output,
  sce,
  sampvar = "Barcode",
  colorvar = "label.main",
  which = c(1, 2),
  nvar = 5,
  shr = 0.6,
  ...
)
}
\arguments{
\item{prcomp_output}{instance of prcomp, as produced by, e.g., irlba::prcomp_irlba}

\item{sce}{a SingleCellExperiment instance}

\item{sampvar}{character(1) name of a variable in `colData(sce)` that will be used to label samples}

\item{colorvar}{character(1) names of a variable in `colData(sce)` that will be used to color points}

\item{which}{numeric(2), gives dimensions of `prcomp_output$x` to use for point locations}

\item{nvar}{numeric(1) number of highly ranked features (by sum of squares of loadings) to use in biplot}

\item{shr}{numeric(1) a fudge factor to 'shrink' projection span relative to boundaries of display}

\item{\dots}{not used}
}
\value{
a ggplot instance
}
\description{
produce a biplot based on sharing of information between a `prcomp`-like output and a SingleCellExperiment
}
