% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doplotly.R
\name{get4kPBMC}
\alias{get4kPBMC}
\title{get 4000+ PBMCs from TENx genomics}
\usage{
get4kPBMC(doNorm = TRUE, useSyms = TRUE)
}
\arguments{
\item{doNorm}{logical(1) if TRUE, apply scater::logNormCounts to PBMC SingleCellExperiment}

\item{useSyms}{logical(1) if TRUE, remove features lacking Symbol in rowData, and use Symbol 
as rownames}
}
\value{
SingleCellExperiment
}
\description{
get 4000+ PBMCs from TENx genomics
}
