% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doplotly.R
\name{getHPCAreference}
\alias{getHPCAreference}
\title{get HPCA reference}
\usage{
getHPCAreference(...)
}
\arguments{
\item{\dots}{passed to celldex::HumanPrimaryCellAtlasData}
}
\value{
SummarizedExperiment
}
\description{
get HPCA reference
}
